chore: extract first image component
I want to use the same 'load first image' controls on the bootable image
empty screen as what we have on the Dashboard today. This just extracts the UI
elements out of the Dashboard and into its own reusable component, which is
better code separation anyway.

Part of #862.
